✨ Feat: Add localization support This commit introduces localization support to the application.  It includes the generation of localization keys, the creation of localization service and the addition of Arabic and English language files.  This allows the app to be translated into multiple languages.
⬆️ Chore: Upgrade compileSdk to 35
This commit upgrades the Android compileSdk version to 35 to ensure compatibility with the latest Android features and security updates.

🔧 Chore: Add mason configuration file
This commit adds a mason.yaml file to the project, enabling the use of Mason for generating project templates and managing project structure.

✅ Test: Add devtools configuration file
This commit adds a devtools_options.yaml file to configure Dart & Flutter DevTools for improved debugging and development workflow.

♻️ Refactor: Improve server.ts port handling
This commit refactors the server.ts file to use an environment variable for the port, improving flexibility and allowing for easier configuration.  The port variable is also renamed to improve semantics.

✨ Feat: Add authentication features
This commit introduces local and remote data sources for authentication, along with a repository and model for managing authentication tokens.  It also includes the implementation of login, password reset, and refresh token functionalities.  This enhances the security and user experience of the application.

✨ Feat: Add core modules
This commit introduces core modules for the application, including constants, injection, models, repositories, and result builders.  This improves the organization and maintainability of the codebase.  It also includes the implementation of a reactive token storage for managing authentication tokens.

✨ Feat: Add root screen and navigation
This commit adds the root screen and navigation for the application.  It includes the implementation of a bottom navigation bar and an end drawer for improved user navigation.  This enhances the user experience and provides a more intuitive way to navigate the application.

📝 Docs: Add comments to generated files
This commit adds comments to the generated localization files to improve readability and understanding.