✨ (l10n): add new localization keys for countdown timer 🐛 (generate_localization_keys.dart): fix import path for localization service 🎨 (core/styles): improve app theme and colors ✨ (home): add countdown timer widget to home screen 🎨 (root): improve root navbar blur effect
Added new localization keys for the countdown timer feature in both Arabic and English. Fixed an incorrect import path in the localization key generator. Improved the app's theme and color scheme for better visual appeal. Integrated a countdown timer widget into the home screen to display the time remaining until the finals. Enhanced the blur effect on the root navigation bar for a smoother user experience.