✨ (grades): Refactor grades feature to use GradesResponseEntity
This commit refactors the grades feature to use a new
`GradesResponseEntity` entity. This entity encapsulates a list of
`GradeEntity` objects, as well as summary information about the grades,
such as the number of passed and failed subjects, and the average grade.

The following changes were made:

- Added new `GradeEntity` and `GradesResponseEntity` entities.
- Updated the `GradesUsecase` to return a `GradesResponseEntity` object.
- Updated the `GradeBloc` to use the `GradesResponseEntity` object.
- Updated the UI to display the summary information from the
  `GradesResponseEntity` object.
- Updated dummy data to use the new entities.

These changes improve the structure and organization of the grades
feature, and make it easier to display summary information about the
grades.