♻️ (announcement_mappers.dart): Refactor AnnouncementMapper to use dummy data and DateTime extension ✨ (announcements_repository_impl.dart): Implement getFacultyAnnouncements and getStudentAnnouncements methods ✨ (announcements_repository.dart): Add getFacultyAnnouncements method ✨ (announcements_usecase.dart): Implement getFacultyAnnouncements use case ✨ (get_my_announcements_usecase.dart): Implement getMyAnnouncement use case 🐛 (announcement_bloc.dart): Fix getAnnouncement method to use getFacultyAnnouncements use case 🐛 (all_news_widget.dart): Fix imageUrl to use dummyImages if image is empty
The AnnouncementMapper extension now uses dummyImages and the fullDate extension for createdAt.
The AnnouncementsRepoImp now implements the getFacultyAnnouncements and getStudentAnnouncements methods.
The AnnouncementsRepo now has a getFacultyAnnouncements method.
The AnnouncementsUsecase now has a getFacultyAnnouncements use case.
The GetMyAnnouncementsUsecase now has a getMyAnnouncement use case.
The AnnouncementBloc now uses the getFacultyAnnouncements use case.
The AllNewsWidget now uses dummyImages if the image is empty.
These changes were made to improve the functionality and readability of the code.