♻️ (root_header.dart): remove notification icon from root header 🐛 (all_news_widget.dart): refetch data on error in all news widget 🐛 (my_news_widget.dart): refetch data on error in my news widget ✨ (home_body.dart): add refresh indicator to home body ♻️ (nav_bar_provider.dart): move nav bar provider to presentation layer ♻️ (nav_bar_provider.dart): change animation to jumpToPage in nav bar provider
The notification icon was removed from the root header as it was not functional and was causing confusion.
The all news and my news widgets now refetch data on error to ensure that the user always sees the most up-to-date information.
The home body now has a refresh indicator to allow the user to manually refresh the data.
The nav bar provider was moved to the presentation layer to improve the project structure.
The animation in the nav bar provider was changed to jumpToPage to improve performance.